style: split prune tests into own file and drop trailing comma
SwiftLint --strict caught:
- KFinderTests.swift exceeded the 400-line file_length limit after
  adding the prune-suite. Moved the suite to its own
  EnforceRecentFoodsLimitTests.swift; the synchronized-group target
  membership picks it up automatically.
- Trailing comma after the new fdcId sort key in
  UserPreferences.enforceRecentFoodsLimit(in:limit:).
